Move views to own folder. Add check for correct answers.
